:truck: Rename partials for distinction
