+++ app/views/_logo.jsx
